+++ straxen/__init__.py
@@ -1,8 +1,9 @@
-__version__ = '0.10.0'
+__version__ = '0.11.0'
 
 from .common import *
 from .itp_map import *
 from .rundb import *
+from .online_monitor import *
 from .matplotlib_utils import *
 from .mini_analysis import *
 from .misc import *