+++ HISTORY.md
@@ -1,3 +1,27 @@
+0.11.0 / 2020-10-15
+--------------------
+- Separate context for fist commissioning data (#246)
+- Online Monitor storage frontend (#216)
+- Add Acquisition-monitor plugins (#207)
+- Many (bug)fixes (#221, #223, #224, #226, #236, #238, #240, #241, #241, #245)
+- Use CutPlugin class (#225)
+- Bootstrax updates (#232)
+
+
+0.10.1 / 2020-09-11
+--------------------
+- Added pytest to travis which builds data for all registered nT plugins (#193)
+- Added moun- and neutron-veto into the streamline processing with bootstrax (#184)
+- Added back-up URLs for the mongoDB (#213)
+- Updated data cleaning/handling with ajax and bootstrax (#182, #191, #196, #202, #206)
+- Updated documentation and doc-strings (#189, #192 and #198)
+- Updated bin scripts like straxer (#204) 
+- Updated PMT gains (#208)
+- Renamed high energy plugins (#200)
+- Bugifx in nveto-plugins (#183, #209)
+- Bugfix in clean_up_empty_records (#210)
+
+
 0.10.0 / 2020-08-187
 --------------------
 - Neutron-veto integration (#86)